Carousel Added
Per customer request carousel images added
Will need to move to Mac add SASS compiler then download
Bootstrap - looking for RFS (responsive font size) which must
be enabled in the SCSS file. Then recompiled. Will be available
as default in Bootstrap 5. Will not be able to use the CDN for
this project.